newly edited

+++ routes/api/posts/posts.js
@@ -31,18 +31,47 @@ router.get("/:id", auth.optional, async (req, res) => {
   }
 });
 
-router.post("/", auth.required, async (req, res, next) => {
-  const result = await postValidator(req.body);
-  if (result.error) return res.status(400).send(error.message);
-
+router.patch("/avatar/:userId", auth.required, async (req, res) => {
   try {
+    const posts = Post.updateMany(
+      { userId: req.params.userId },
+      { avatar: req.body.avatar }
+    );
+
+    const comments = Post.updateMany(
+      {},
+      {
+        $set: { "comments.$[elem].avatar": req.body.avatar }
+      },
+      { arrayFilters: [{ "elem.userId": req.params.userId }] }
+    );
+
+    const result = await Promise.all([posts, comments]);
+
+    return res.json({
+      postModified: result[0].nModified,
+      commentsModified: result[1].nModified
+    });
+  } catch (error) {
+    return res.status(500);
+  }
+});
+
+router.post("/", auth.required, async (req, res, next) => {
+  try {
+    const result = await postValidator(req.body);
+    if (result.error) return res.status(400).send(error.message);
+
     //save to db
     const dbSchema = {
       title: result.title,
       author: result.author,
       content: result.content,
       tags: result.tags ? result.tags : [],
-      likes: result.likes ? result.likes : 0
+      likes: result.likes ? result.likes : 0,
+      post_date_timestamp: new Date().getTime(),
+      userId: result.userId,
+      avatar: result.avatar
     };
 
     let post = new Post(dbSchema);
@@ -55,41 +84,42 @@ router.post("/", auth.required, async (req, res, next) => {
 
     res.status(201).json(post);
   } catch (error) {
-    return res.status(400).json(error.message);
+    console.log(error);
+    return res.status(412).json(error.message);
   }
 });
 
-router.delete("/:id", (req, res) => {
-  const promiseAlgolia = index.deleteObject(req.params.id);
+router.delete("/:id", async (req, res) => {
+  try {
+    const result = await index.deleteObject(req.params.id);
 
-  const promiseMongo = Post.findOneAndDelete({ _id: req.params.id });
+    if (!result) return res.status(400);
 
-  Promise.all([promiseAlgolia, promiseMongo])
-    .then(content => res.json(content))
-    .catch(err => res.status(500).json(err.message));
+    await Post.findOneAndDelete({ _id: req.params.id });
+
+    res.json({ message: "deleted" });
+  } catch (error) {
+    res.status(500);
+  }
 });
 
 router.patch("/likes/:id", auth.required, async (req, res) => {
   const increment = req.body.liked ? -1 : 1;
 
   try {
-    const content = await index.getObject(req.params.id, ["likes"]);
+    // const content = await index.getObject(req.params.id, ["likes"]);
 
+    const content = await Post.findOneAndUpdate(
+      { _id: req.params.id },
+      { $inc: { likes: increment } },
+      { new: true }
+    );
     //Because the aync feature of algolia
     //We have to waittask for its update to keep consistency
-    await index.partialUpdateObject(
-      {
-        likes: content.likes + increment,
-        objectID: req.params.id
-      },
-      (err, { taskID } = {}) => {
-        index.waitTask(taskID, err => {
-          if (!err) {
-            return res.json({ message: "updated" });
-          }
-        });
-      }
-    );
+    await index.partialUpdateObject({
+      likes: content.likes,
+      objectID: req.params.id
+    });
   } catch (error) {
     return res.json({ message: error.message });
   }
@@ -117,6 +147,25 @@ router.patch("/comments/:id", auth.required, async (req, res) => {
 
     return res.json(comment);
   } catch (error) {
+    console.log(error);
+    return res.status(403).send(error.message);
+  }
+});
+
+router.delete("/comments/:id", async (req, res) => {
+  try {
+    const post = await Post.findOneAndUpdate(
+      { _id: req.params.id },
+      {
+        $pull: { comments: { _id: req.query.commentId } }
+      },
+      { new: true }
+    );
+
+    if (!post) return res.status(400).send("Post not Found");
+
+    return res.json(post);
+  } catch (error) {
     return res.status(403).send(error.message);
   }
 });
@@ -138,10 +187,10 @@ router.patch("/tags/:id", auth.required, async (req, res) => {
 });
 
 router.patch("/comments/reply/:id", auth.required, async (req, res) => {
-  const { error } = await replyValidator(req.body.reply);
-  if (error) return res.status(400).send(error.message);
-
   try {
+    const { error } = await replyValidator(req.body.reply);
+    if (error) return res.status(400).send(error.message);
+
     let post = await Post.findOne({ _id: req.params.id });
 
     let result = await post.comments.id(req.query.commentId);
@@ -152,6 +201,22 @@ router.patch("/comments/reply/:id", auth.required, async (req, res) => {
     post = await post.save();
 
     res.json(newReply);
+  } catch (error) {
+    res.json(error.message);
+  }
+});
+
+router.delete("/comments/reply/:postId", auth.required, async (req, res) => {
+  try {
+    let post = await Post.findOne({ _id: req.params.postId });
+
+    let result = await post.comments.id(req.query.commentId);
+
+    await result.replies.id(req.query.replyId).remove();
+
+    post = await post.save();
+
+    res.json(post);
   } catch (error) {
     res.json(error.message);
   }
